Composer create project

+++ scripts/host/m_switch_to_ee.sh
@@ -11,46 +11,53 @@ magento_ce_dir="${vagrant_dir}/magento2ce"
 magento_ee_dir="${magento_ce_dir}/magento2ee"
 host_os="$(bash "${vagrant_dir}/scripts/host/get_host_os.sh")"
 php_executable="$(bash "${vagrant_dir}/scripts/host/get_path_to_php.sh")"
+checkout_source_from="$(bash "${vagrant_dir}/scripts/get_config_value.sh" "checkout_source_from")"
 
-force_switch=0
-while getopts 'f' flag; do
-  case "${flag}" in
-    f) force_switch=1 ;;
-    *) error "Unexpected option" && decrementNestingLevel && exit 1;;
-  esac
-done
+if [[ "${checkout_source_from}" == "git" ]]; then
+    # Current installation is Git-based
+    force_switch=0
+    while getopts 'f' flag; do
+      case "${flag}" in
+        f) force_switch=1 ;;
+        *) error "Unexpected option" && decrementNestingLevel && exit 1;;
+      esac
+    done
 
-if [[ ! -f ${magento_ee_dir}/LICENSE_EE.txt ]]; then
-    error "EE codebase is not available."
-    decrementNestingLevel
-    exit 0
-else
-    if [[ -f ${magento_ce_dir}/LICENSE_EE.txt ]] && [[ ${force_switch} -eq 0 ]]; then
-        warning "Already switched to EE. Use 'm-switch-to-ee -f' to switch anyway."
+    if [[ ! -f ${magento_ee_dir}/LICENSE_EE.txt ]]; then
+        error "EE codebase is not available."
         decrementNestingLevel
         exit 0
+    else
+        if [[ -f ${magento_ce_dir}/LICENSE_EE.txt ]] && [[ ${force_switch} -eq 0 ]]; then
+            warning "Already switched to EE. Use 'm-switch-to-ee -f' to switch anyway."
+            decrementNestingLevel
+            exit 0
+        fi
+
+        status "Linking EE repository"
+        ${php_executable} -f ${magento_ee_dir}/dev/tools/build-ee.php -- --command=link --ee-source="${magento_ee_dir}" --ce-source="${magento_ce_dir}" --exclude=true 2> >(logError) > >(log)
+
+        cp ${magento_ee_dir}/composer.lock ${magento_ce_dir}/composer.lock
+
+        if [[ ${host_os} == "Windows" ]] || [[ $(bash "${vagrant_dir}/scripts/get_config_value.sh" "guest_use_nfs") == 0 ]]; then
+            # Prevent issues on Windows with incorrect symlinks to files
+            if [[ -f ${magento_ee_dir}/app/etc/aliases_to_classes_map.json ]] && [[ -L ${magento_ce_dir}/app/etc/aliases_to_classes_map.json ]]; then
+                rm ${magento_ce_dir}/app/etc/aliases_to_classes_map.json
+                cp ${magento_ee_dir}/app/etc/aliases_to_classes_map.json ${magento_ce_dir}/app/etc/aliases_to_classes_map.json
+            fi
+            if [[ -f ${magento_ee_dir}/LICENSE_EE.txt ]] && [[ -L ${magento_ce_dir}/app/etc/enterprise ]]; then
+                rm ${magento_ce_dir}/app/etc/enterprise
+                mkdir ${magento_ce_dir}/app/etc/enterprise
+                cp ${magento_ee_dir}/LICENSE_EE.txt ${magento_ce_dir}/LICENSE_EE.txt
+            fi
+        fi
     fi
 
-    status "Linking EE repository"
-    ${php_executable} -f ${magento_ee_dir}/dev/tools/build-ee.php -- --command=link --ee-source="${magento_ee_dir}" --ce-source="${magento_ce_dir}" --exclude=true 2> >(logError) > >(log)
-
-    cp ${magento_ee_dir}/composer.lock ${magento_ce_dir}/composer.lock
-
-    if [[ ${host_os} == "Windows" ]] || [[ $(bash "${vagrant_dir}/scripts/get_config_value.sh" "guest_use_nfs") == 0 ]]; then
-        # Prevent issues on Windows with incorrect symlinks to files
-        if [[ -f ${magento_ee_dir}/app/etc/aliases_to_classes_map.json ]] && [[ -L ${magento_ce_dir}/app/etc/aliases_to_classes_map.json ]]; then
-            rm ${magento_ce_dir}/app/etc/aliases_to_classes_map.json
-            cp ${magento_ee_dir}/app/etc/aliases_to_classes_map.json ${magento_ce_dir}/app/etc/aliases_to_classes_map.json
-        fi
-        if [[ -f ${magento_ee_dir}/LICENSE_EE.txt ]] && [[ -L ${magento_ce_dir}/app/etc/enterprise ]]; then
-            rm ${magento_ce_dir}/app/etc/enterprise
-            mkdir ${magento_ce_dir}/app/etc/enterprise
-            cp ${magento_ee_dir}/LICENSE_EE.txt ${magento_ce_dir}/LICENSE_EE.txt
-        fi
-    fi
+    bash "${vagrant_dir}/scripts/host/relink_sample_data.sh" 2> >(logError)
+else
+    # Current installation is Composer-based
+    warning "Switching between CE and EE is not possible for composer-based installation. Falling back to reinstall"
 fi
-
-bash "${vagrant_dir}/scripts/host/relink_sample_data.sh" 2> >(logError)
 
 bash "${vagrant_dir}/scripts/host/m_clear_cache.sh" 2> >(logError)
 bash "${vagrant_dir}/scripts/host/m_composer.sh" install 2> >(logError)
